requirements.txt
pytest==9.0.2
playwright==1.58.0
pytest-playwright==0.7.2
allure-python-commons
pydantic
pydantic-settings